fix: cleanup tests
